Hoist array-move
