+++ packages/user/FractalCore/ui/src/pages/fractal/evaluations/evaluation-deliberation.tsx
@@ -12,8 +12,8 @@ import { useWatchClose } from "@/hooks/fractals/use-watch-close";
 import { useEvaluationStatus } from "@/hooks/use-evaluation-status";
 import { useGuildAccount } from "@/hooks/use-guild-account";
 import { useNowUnix } from "@/hooks/use-now-unix";
-import { arrayMove } from "@/lib/arrayMove";
 
+import { arrayMove } from "@shared/lib/array-move";
 import { humanize } from "@shared/lib/humanize";
 import { Account } from "@shared/lib/schemas/account";
 import { Badge } from "@shared/shadcn/ui/badge";
